refactor: use class resolver for Dataloaders init
We are currently relying on the class resolver to instantiate "flat"
objects and plug them together later on. However this was not done
before so we also had to extend the class resolver's ability to init
nested objects with "extra_kwargs".

Also the resolver looks now for nested configs to use. Beforehand it
searched only the second level of the config and stopped after
encountering the first "type_hint" on a layer without
further diving.