refactor!: :recycle: lay groundwork for packed datasets
This commit lays the groundwork for groundwork for possible future
packing variants. It adapts the packed index creation in such a way so
that it outputs a binary file which is strucured according to the
following format:

header segment | data segment | index segment

The header segment contains a single 4 bytes integer denoting the
length (in bytes) of the data segment.
The data segment contains a concatenation of the tokenized documents,
seperated by an eos token. Each token is a 4 bytes integer.
The index segment contains a pickled index into the data segment. The
index is a list of tuples, where every tuple contains the offset and the
length (in bytes) of a document.

Furthermore, PackedMemMapDatasetBase was introduced as a base
class which loads and parses the packed files. The naive packed dataset
implementation was moved to PackedMemMapDatasetContinuous.